feat: establish shared web UI lifecycle and primitives
