+++ package.json
@@ -72,6 +72,7 @@
     "@dnd-kit/utilities": "^3.2.2",
     "@fastify/cors": "^11.2.0",
     "@fastify/static": "^9.0.0",
+    "@radix-ui/react-dialog": "^1.1.23",
     "@visactor/react-vchart": "^2.0.22",
     "better-sqlite3": "^12.10.0",
     "dotenv": "^17.4.2",
