+++ src/web/components/ResponsiveBatchActionBar.architecture.test.ts
@@ -10,7 +10,7 @@ describe('ResponsiveBatchActionBar page adoption', () => {
   it('is used by the repeated admin list pages instead of open-coded mobile and desktop batch wrappers', () => {
     expect(readPageSource('src/web/pages/Accounts.tsx')).toContain('ResponsiveBatchActionBar');
     expect(readPageSource('src/web/pages/Sites.tsx')).toContain('ResponsiveBatchActionBar');
-    expect(readPageSource('src/web/pages/Tokens.tsx')).toContain('ResponsiveBatchActionBar');
+    expect(readPageSource('src/web/pages/tokens/TokensPanel.tsx')).toContain('ResponsiveBatchActionBar');
     expect(readPageSource('src/web/pages/DownstreamKeys.tsx')).toContain('ResponsiveBatchActionBar');
   });
 });
