+++ src/web/components/ui/Drawer.tsx
@@ -1,0 +1,5 @@
+export {
+  Drawer as default,
+  Drawer,
+  type DrawerProps,
+} from './Dialog.js';
