+++ src/web/pages/tokens.actions-layout.test.ts
@@ -4,7 +4,7 @@ import { resolve } from 'node:path';
 
 describe('Tokens actions layout', () => {
   it('reuses a wrapping actions cell layout so token row actions do not overflow', () => {
-    const source = readFileSync(resolve(process.cwd(), 'src/web/pages/Tokens.tsx'), 'utf8');
+    const source = readFileSync(resolve(process.cwd(), 'src/web/pages/tokens/TokensPanel.tsx'), 'utf8');
     const css = readFileSync(resolve(process.cwd(), 'src/web/index.css'), 'utf8');
 
     expect(source).toContain('className="token-actions-cell"');
