+++ src/web/pages/tokens.embedded-layout.test.ts
@@ -4,7 +4,7 @@ import { resolve } from 'node:path';
 
 describe('Tokens embedded layout', () => {
   it('reuses accounts-page-actions layout when embedded in 连接管理', () => {
-    const source = readFileSync(resolve(process.cwd(), 'src/web/pages/Tokens.tsx'), 'utf8');
+    const source = readFileSync(resolve(process.cwd(), 'src/web/pages/tokens/TokensPanel.tsx'), 'utf8');
 
     expect(source).toContain("className={`page-actions ${embedded ? 'accounts-page-actions' : ''}`.trim()}");
   });
